requirements.txt
anyio==4.9.0
apify_client==1.10.0
apify_shared==1.4.1
certifi==2025.4.26
charset-normalizer==3.4.1
dnspython==2.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
more-itertools==10.7.0
pymongo==4.12.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
urllib3==2.4.0
